+++ gwcelery/tests/test_tools_flask.py
@@ -11,6 +11,7 @@ def test_flask_run(monkeypatch):
     mock_run_simple = Mock()
     monkeypatch.setattr('werkzeug.serving.run_simple', mock_run_simple)
     monkeypatch.setenv('FLASK_PORT', '5556')
+    monkeypatch.setattr(celery_app.log, 'setup', Mock())
 
     with pytest.raises(SystemExit) as excinfo:
         celery_app.start(['gwcelery', 'flask', 'run', '--eager-loading'])