+++ gwcelery/tasks/external_triggers.py
@@ -230,19 +230,15 @@ def handle_grb_lvalert(alert):
             gw_group = gracedb.get_group(preferred_event_id)
             raven.coincidence_search(graceid, alert['object'],
                                      group=gw_group, searches=['GRB'])
-            if gw_group == 'CBC':
-                subthresh_search = True
-            else:
-                subthresh_search = False
+            subthresh_search = gw_group == 'CBC'
         if subthresh_search:
             # launch subthreshold searches if CBC or subthreshold GRB
             # for Fermi and Swift separately to use different time windows
-            raven.coincidence_search(graceid, alert['object'], group='CBC',
-                                     searches=['SubGRB', 'SubGRBTargeted'],
-                                     pipelines=['Fermi'])
-            raven.coincidence_search(graceid, alert['object'], group='CBC',
-                                     searches=['SubGRB', 'SubGRBTargeted'],
-                                     pipelines=['Swift'])
+            for pipeline in ['Fermi', 'Swift']:
+                raven.coincidence_search(
+                    graceid, alert['object'], group='CBC',
+                    searches=['SubGRB', 'SubGRBTargeted'],
+                    pipelines=[pipeline])
     elif alert['alert_type'] == 'label_added' and \
             alert['object'].get('group') == 'External':
         if _skymaps_are_ready(alert['object'], alert['data']['name'],