+++ CHANGES.rst
@@ -18,9 +18,10 @@ Changelog
 -   Append to, and do not overwrite, log files, when starting GWCelery via
     Condor.
 
--   Launch raven coincidence search for subthreshold GRBs separately in order
-    to use different time windows. This enables the joint LVK-Fermi and
-    LVK-Swift targeted searches to b integrated with RAVEN.
+-   Launch raven coincidence search for subthreshold GRBs separately for
+    different gamma-ray expermients in order to use different time windows.
+    This enables the joint LVK-Fermi and LVK-Swift targeted searches to be
+    integrated with RAVEN.
 
 -   Grab subGRB Fermi sky maps from GCN.