+++ CHANGES.rst
@@ -19,8 +19,8 @@ Changelog
 -   Append to, and do not overwrite, log files, when starting GWCelery via
     Condor.
 
--   Launch raven coincidence search for subthreshold GRBs separately for
-    different gamma-ray expermients in order to use different time windows.
+-   Launch raven coincidence search for sub-threshold GRBs separately for
+    different gamma-ray experiments in order to use different time windows.
     This enables the joint LVK-Fermi and LVK-Swift targeted searches to be
     integrated with RAVEN.